% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JhPurchase.R
\name{purchaseContract_update}
\alias{purchaseContract_update}
\title{反写采购合同}
\usage{
purchaseContract_update(token)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}

\item{FName}{ERP账套名}
}
\value{

}
\description{
反写采购合同
}
\examples{
purchaseContract_update('C0426D23-1927-4314-8736-A74B2EF7A039')
}
